Move terminalInstance, electron-browser->browser
Part of #69115
