+++ src/tsconfig.strictNullChecks.json
@@ -282,6 +282,7 @@
 		"./vs/workbench/contrib/terminal/browser/terminalCommandTracker.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalConfigHelper.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalFindWidget.ts",
+		"./vs/workbench/contrib/terminal/browser/terminalInstance.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalLinkHandler.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalPanel.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalQuickOpen.ts",
@@ -293,7 +294,6 @@
 		"./vs/workbench/contrib/terminal/common/terminalEnvironment.ts",
 		"./vs/workbench/contrib/terminal/common/terminalMenu.ts",
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
-		"./vs/workbench/contrib/terminal/electron-browser/terminalInstance.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcessExtHostProxy.ts",
