+++ src/vs/workbench/contrib/terminal/browser/terminalInstance.ts
